Getting Git repo to reflect live website exactly.
Removed the README.md file and added the favicon.ico file. The idea here
is to get the repo to reflect the live website. So, you can pull a copy
of the master branch repo files and install to the web server with all
the needed files and structure and nothing more than what is needed.