December 12, 2017 web edits.
Fix broken link to FAQ in GitHub. There are some extra files getting
committed here. Per the web developer;

"Everything there is exactly what is on staging. Those extra untracked
files are innocuous. Safe to add them, but they’ll likely be removed
with the next push. They’re not referenced by code at the moment."

So, in the interest of reflecting what is on the web site these are
being added so that we know what is on the site as of now.

 Changes to be committed:
	modified:   css/styles.min.css
	modified:   features/index.html
	modified:   index.html
	modified:   js/lib/jquery.countdown.min.js
	modified:   js/lib/particles.min.js
	modified:   js/main.js
	modified:   media-kit/index.html
	modified:   press/index.html
	modified:   privacy/index.html

 Untracked files:
	android-chrome-192x192.png
	apple-touch-icon.png
	browserconfig.xml
	favicon-16x16.png
	favicon-32x32.png
	favicon.png
	images/bg-featues-container.png
	manifest.json
	safari-pinned-tab.svg
	site.webmanifest
	touch-icon.png